Фикс слешей + команда на отправку дейли команды

+++ la_rzn_bot.py
@@ -45,18 +45,22 @@ async def send_message(message: types.Message):
     params.set_last_message_id(0)
 
 
-# @dp.message(Command("daily"))
+@dp.message(Command("daily"))
 async def daily(message: types.Message):
-    message_id = params.get_last_message_id()
-    if message_id != 0:
-        await bot.delete_message(chat_id=params.la_chat_id, message_id=message_id)
-    sent_message = await bot.send_photo(chat_id=params.la_chat_id, message_thread_id=params.bk_thread_id,
-                                        photo=types.FSInputFile("resources\\qr.png"),
-                                        caption=params.generate_daily_message(), parse_mode="HTML",
-                                        reply_to_message_id=params.get_reply_message_id(),
-                                        reply_markup=keyboard.get_vote_button_keyboard())
-    log(f"ID отправленного сообщения: {sent_message.message_id}")
-    params.set_last_message_id(sent_message.message_id)
+    if message.chat.id == 649062985:
+        try:
+            message_id = params.get_last_message_id()
+            if message_id != 0:
+                await bot.delete_message(chat_id=params.la_chat_id, message_id=message_id)
+            sent_message = await bot.send_photo(chat_id=params.la_chat_id, message_thread_id=params.bk_thread_id,
+                                                photo=types.FSInputFile("resources\\qr.png"),
+                                                caption=params.generate_daily_message(), parse_mode="HTML",
+                                                reply_to_message_id=params.get_reply_message_id(),
+                                                reply_markup=keyboard.get_vote_button_keyboard())
+            log(f"ID отправленного сообщения: {sent_message.message_id}")
+            params.set_last_message_id(sent_message.message_id)
+        except Exception as e:
+            log(f"Ошибка при отправке сообщения: {e}")
 
 
 @dp.message()
